Module name reassignments adapted to epydoc
Scientific.Geometry and Scientific.BSP contain some objects that are
implemented in submodules that should not be documented because they
are implementation details that might change in the future. Their
module names are therefore manipulated when imported for documentation
generation (tested as the presence of epydoc in sys.modules). The
formerly used mechanism for pydoc needed some updating, which is
now done.