DistributedComputing: made the watchdog period of slave processes a user-definable parameter
The watchdog signal is sent by a separate thread, but depending on what
executes in the main thread (worst case: a long calculation that does
not release the Global Interpreter Lock) it cannot guarantee its
timing. In that case, the task manager may write off a slave process
as dead although it is really hard at work. With this modification,
users can at least work around this problem by increasing the
watchdog period or by setting it to None.